feat(dtfs2-7075): update validation
