+++ libs/common/src/helpers/payload-verification/is-verified-payload-by-type.ts
@@ -1,10 +1,10 @@
-import { IsVerifiedPayloadbyTypeParams, TemplateByType } from '../../types/payload-verification';
+import { IsVerifiedPayloadByTypeParams } from '../../types/payload-verification';
 
 export const isVerifiedPayloadByType = ({
   payload,
   template,
   areAllPropertiesRequired,
-}: IsVerifiedPayloadbyTypeParams) => {
+}: IsVerifiedPayloadByTypeParams) => {
   if (!payload || !template || !Object.keys(payload).length || !Object.keys(template).length) {
     return false;
   }
@@ -13,10 +13,17 @@ export const isVerifiedPayloadByType = ({
   const payloadKeys = Object.keys(payload);
   const templateKeys = Object.keys(template);
 
-  const doesEveryTemplatePropertyExistOnThePayload = templateKeys.every((propertyKey) => payloadKeys.includes(propertyKey));
+  const doesEveryTemplatePropertyExistOnThePayload = templateKeys.every((propertyKey) =>
+    payloadKeys.includes(propertyKey),
+  );
+
+  const doesEveryTemplatePropertyExistOnThePayloadIfRequired =
+    !areAllPropertiesRequired || doesEveryTemplatePropertyExistOnThePayload;
 
   // 2. Ensure no additional properties
-  const doesthePayloadHaveAnyAdditionalProperties = !payloadKeys.filter((propertyKey) => !templateKeys.includes(propertyKey)).length;
+  const doesthePayloadHaveAnyAdditionalProperties = !payloadKeys.filter(
+    (propertyKey) => !templateKeys.includes(propertyKey),
+  ).length;
 
   // 3. Properties data type validation
   const doesTypesMatch = payloadKeys.every((key) => {
@@ -27,5 +34,7 @@ export const isVerifiedPayloadByType = ({
   });
 
   // Compound comparison condition
-  return doesEveryTemplatePropertyExistOnThePayload && doesthePayloadHaveAnyAdditionalProperties && doesTypesMatch;
+  return (
+    doesEveryTemplatePropertyExistOnThePayloadIfRequired && doesthePayloadHaveAnyAdditionalProperties && doesTypesMatch
+  );
 };
