+++ libs/common/src/types/payload-verification.ts
@@ -1,14 +1,14 @@
 import z from 'zod';
 import { Prettify } from './types-helper';
 
-export type TemplateByType = Record<string, string>;
-
 type BaseIsVerifiedPayloadParams = {
   payload: Record<string, unknown>;
   areAllPropertiesRequired?: boolean;
 };
-export type IsVerifiedPayloadbyZodParams = Prettify<BaseIsVerifiedPayloadParams & { template: z.AnyZodObject }>;
+export type IsVerifiedPayloadByZodParams = Prettify<BaseIsVerifiedPayloadParams & { template: z.AnyZodObject }>;
 
-export type IsVerifiedPayloadbyTypeParams = Prettify<BaseIsVerifiedPayloadParams & { template: TemplateByType }>;
+export type IsVerifiedPayloadByTypeParams = Prettify<
+  BaseIsVerifiedPayloadParams & { template: Record<string, string> }
+>;
 
-export type IsVerifiedPayloadParams = IsVerifiedPayloadbyZodParams | IsVerifiedPayloadbyTypeParams;
+export type IsVerifiedPayloadParams = IsVerifiedPayloadByZodParams | IsVerifiedPayloadByTypeParams;
